Rename other modules according to pep8
