+++ reskit/csp/workflows/__init__.py
@@ -1,3 +1,3 @@
-from . import csp_workflow_manager
-from . import workflows
-from . import dataset_handler
+from reskit.csp.workflows import csp_workflow_manager
+from reskit.csp.workflows import workflows
+from reskit.csp.workflows import dataset_handler
