--- reskit/weather/IconlamSource/__init__.py
@@ -1,1 +1,0 @@
-from .IconlamSource import IconlamSource
