+++ reskit/weather/__init__.py
@@ -7,7 +7,7 @@ from reskit.weather.era_5_source.era_5_source import Era5Source
 from reskit.weather.sarah_source import SarahSource
 from reskit.weather.gsa_mean.gsa_mean_source import GSAmeanSource
 from reskit.weather.gwa_mean.gwa_mean_source import GWAmeanSource
-from reskit.weather.IconlamSource.IconlamSource import IconlamSource
+from reskit.weather.iconlam_source.iconlam_source import IconlamSource
 
 # from .CosmoSource import CosmoSource  # TODO: NEEDS UPDATING!
 # from .CordexSource import CordexSourc  # TODO: NEEDS UPDATING!
