+++ reskit/weather/iconlam_source/__init__.py
@@ -1,0 +1,1 @@
+from reskit.weather.iconlam_source.iconlam_source import IconlamSource
